Revert "GCC4.6: Convert various empty macros to inline functions"
This reverts commit e3bd6eb1448aa87873d5c670fafe803033da5341.

The commit causes build breakage for a number of boards. This results
from the fact that now the arguments of debug() actually get
referenced (even if there is hope that the compiler will optimize
away the debug() call).  The obvious fix to that probem (change the
code to always declare the referenced variables and data structures)
increases the code size, and was this rejected.  So it was decided to
revert this commit until a better solution is found.